requirements.txt
# 深度学习框架
torch>=2.0.0
torch-geometric>=2.3.0
torchvision>=0.15.0

# 数据处理
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0

# 可视化
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.15.0

# 开发工具
pytest>=7.4.0
black>=23.0.0
mypy>=1.5.0
flake8>=6.0.0
pre-commit>=3.3.0

# 工具库
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Jupyter支持
jupyter>=1.0.0
ipykernel>=6.25.0

# 类型检查支持
types-PyYAML>=6.0.0
types-tqdm>=4.65.0

# 数据科学工具
scipy>=1.10.0
networkx>=3.1.0

# 配置管理
hydra-core>=1.3.0
omegaconf>=2.3.0

# 实验跟踪（可选）
tensorboard>=2.13.0
wandb>=0.15.0

# 性能分析
psutil>=5.9.0
memory-profiler>=0.61.0
